+++ jest.config.js
@@ -3,7 +3,8 @@ export default {
   preset: 'ts-jest',
   testEnvironment: 'node',
   transform: {
-    '^.+\\.tsx?$': 'ts-jest'
+    '^.+\\.tsx?$': ['ts-jest', { useESM: true }],
+    '^.+\\.js$': 'babel-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js'],
   testMatch: ['**/?(*.)+(spec|test).[tj]s?(x)'],